6. Zigzag Conversion and unit tests refactoring

+++ Problems/Hard/0004/findMedianSortedArrays.cpp
@@ -12,5 +12,5 @@ double Solution::findMedianSortedArrays(std::vector<int>& nums1, std::vector<int
 {
    nums1.insert(nums1.begin(), nums2.begin(), nums2.end());
    std::sort(nums1.begin(), nums1.end());
-   return (nums1.size() % 2 == 0) ? (nums1[nums1.size() / 2] + nums1[nums1.size() / 2 - 1]) / 2.0 : nums1[nums1.size() / 2];
+   return (nums1.size() % 2 == 0) ? (static_cast<double>(nums1[nums1.size() / 2]) + static_cast<double>(nums1[nums1.size() / 2 - 1])) / 2.0 : nums1[nums1.size() / 2];
 }
